requirements.txt
requests
python-dotenv
PyPDF2
google-genai
fastapi
uvicorn
pydantic
SQLAlchemy
pytest